RimuRealTime

Documentation for RimuRealTime.

RimuRealTime.RimuRealTime

RimuRealTime.Clock

RimuRealTime.ClockAddress

RimuRealTime.ClockObservable

RimuRealTime.ClockOperator

RimuRealTime.ConstantScaling

RimuRealTime.DiscretizedEvolution

RimuRealTime.DynamicScaling

RimuRealTime.Euler

RimuRealTime.EulerSingleState

RimuRealTime.EvolutionStrategy

RimuRealTime.ExponentialSampler

RimuRealTime.FirstOrderTimeEvolution

RimuRealTime.NoScaling

RimuRealTime.PEC

RimuRealTime.PECSingleState

RimuRealTime.Product

RimuRealTime.ProductSingleState

RimuRealTime.QDAlgorithm

RimuRealTime.QDReplicaState

RimuRealTime.QDSimulation

RimuRealTime.QDSimulationPlan

RimuRealTime.QDSingleState

RimuRealTime.QuantumDynamicsProblem

RimuRealTime.RKSingleState

RimuRealTime.RungeKutta

RimuRealTime.ScalingStrategy

RimuRealTime.TimeStepParameters

RimuRealTime.WalkerControl

CommonSolve.init

CommonSolve.solve

CommonSolve.solve!

CommonSolve.step!

RimuRealTime.NthOrderTimeEvolution

RimuRealTime.advance!

RimuRealTime.clock_projector

RimuRealTime.fock_address

RimuRealTime.num_steps

RimuRealTime.scale_state_vector!

RimuRealTime.starting_state

RimuRealTime.time_evolution_operator

RimuRealTime.time_index

RimuRealTime.RimuRealTime — Module

RimuRealTime

RimuRealTime is a package for simulating many-body quantum systems in real time.

Welcome to RimuRealTime version 0.1.0-DEV !

RimuRealTime.Clock — Type

Clock(u, length; start_at, penalty=1.0) <: AbstractHamiltonian{ComplexF64}

Clock Hamiltonian using time evolution operator u with length time steps. The operator u should be a unitary or approximately unitary AbstractOperator with a defined adjoint; for example FirstOrderTimeEvolution, NthOrderTimeEvolution, or ExponentialSampler.

The optional argument start_at specifies the state of the system at $t=0$, otherwise this defaults to DVec(starting_address(u) => 1.0). This should be an AbstractDVec in the underlying Fock space.

penalty is the size of the multiplier on the $t=0$ diagonal element that forces the initial state to remain as start_at.

The address type for the Clock Hamiltonian is ClockAddress - these addresses are built from addresses in the underlying Fock space and an additional time index, which runs from 0 to length.

Information about the clock is accessed with the following:

time_evolution_operator(clock) - the time evolution operator u

num_steps(clock) - the number of time steps length

starting_state(clock) - the starting state start_at

For calculating observables with Clock Hamiltonians, see ClockObservable, ClockOperator, clock_projector.

RimuRealTime.ClockAddress — Type

ClockAddress(address, t) <: AbstractFockAddress

Address type for use with Clocks. Stores an address in the underlying Fock space, and a time step index t. These are accessed with fock_address and time_index.

RimuRealTime.ClockObservable — Type

ClockObservable(obs::AbstractObservable, t::Int) <: AbstractObservable

Observable wrapper for use in replica strategy AllOverlaps with Clock Hamiltonians. Observable expectation values at time index t are calculated using a ClockObservable for the numerator and a clock_projector with the same time index for the denominator. Where possible, a ClockOperator is returned instead of a ClockObservable for efficiency.

RimuRealTime.ClockOperator — Type

ClockOperator(op::AbstractOperator, t::Int) <: AbstractOperator

Operator wrapper for use in replica strategy AllOverlaps with Clock Hamiltonians. Operator expectation values at time index t are calculated using a ClockOperator for the numerator and a clock_projector with the same time index for the denominator.

RimuRealTime.ConstantScaling — Type

ConstantScaling(scale) <: ScalingStrategy

Scale the vector by $\exp(-scale*dt)$ every step. The exponential is approximated by the EvolutionStrategy.

RimuRealTime.DiscretizedEvolution — Type

DiscretizedEvolution(; kwargs...) <: QDAlgorithm

Algorithm used to solve a QuantumDynamicsProblem by evolving a state in time using an approximation to the exponential time evolution operator over many steps. Whether the evolution is deterministic or stochastic is controlled by the keyword arguments style or start_at passed to QuantumDynamicsProblem.

Keyword arguments:

time_step_strategy = ConstantTimeStep(): How to update the time step.

evolution_strategy = PEC(): Which EvolutionStrategy to use to evolve the state.

scaling_strategy = NoScaling(): Which ScalingStrategy to use to control the walker number.

RimuRealTime.DynamicScaling — Type

DynamicScaling(target_walkers) <: ScalingStrategy

Scale the vector every step so that the 1-norm is equal to target_walkers. The cumulative scale is stored in the DataFrame as scale.

RimuRealTime.Euler — Type

Euler() <: EvolutionStrategy

EvolutionStrategy for evolution using the first-order Euler method. In each step the state is updated according to $v_{n+1} = (1 - i H dt)v_n$.

RimuRealTime.EulerSingleState — Type

EulerSingleState(v, wm, id, hamiltonian, time_step) <: QDSingleState

Struct holding state vector and other vectors required for Euler time evolution. See QDReplicaState.

RimuRealTime.EvolutionStrategy — Type

EvolutionStrategy

Abstract type for time evolution strategies. Passed as a parameter to QuantumDynamicsProblem or to DiscretizedEvolution.

Implemented strategies:

PEC

RungeKutta

Euler

Product

RimuRealTime.ExponentialSampler — Type

ExponentialSampler(H<:AbstractHamiltonian, coeff::Number) <: AbstractHamiltonian

Hamiltonian that samples exp(coeff*H). This Hamiltonian does not have iterable offdiagonals, so exact spawning is not possible.

RimuRealTime.FirstOrderTimeEvolution — Type

FirstOrderTimeEvolution(H::AbstractHamiltonian, dt) <: AbstractOperator{ComplexF64}

Time evolution operator that approximates time evolution under a Hamiltonian H as $\exp(-iHdt) \approx 1 - iHdt$. Apply to an AbstractDVec using Rimu.apply_operator! to evolve the state.

RimuRealTime.NoScaling — Type

NoScaling() <: ScalingStrategy

Default ScalingStrategy that does not scale the vector.

RimuRealTime.PEC — Type

PEC(damping=0) <: EvolutionStrategy

EvolutionStrategy for evolution using a second-order predict-evaluate-correct algorithm. This requires only one application of the Hamiltonian per time step. The state is updated every time step according to $v_{n+1} = v_n - i \frac{dt}{2}((1 - d) x_n + (1 + d) x_{n+1})$, where $x_{n+1} = H w_{n+1}$, with $w_{n+1} = v_n - idt x_n$. The vector $x$ is initialized as $x_0 = H v_0$. $d$ is the damping coefficient that modifies the second-order term. Second-order damping can counteract the effects of large spectral components in the Hamiltonian that may lead to an unphysical growth of the 2-norm of the state vector.

RimuRealTime.PECSingleState — Type

PECSingleState(v, working_memory, id, hamiltonian, shift, damping) <: QDSingleState

Struct holding state vector and other vectors required for PEC time evolution. See QDReplicaState.

RimuRealTime.Product — Type

Product(n) <: EvolutionStrategy

EvolutionStrategy for evolution using an nth order expansion of the exponential time evolution operator $\exp(-i H dt)$, where powers of the Hamiltonian are applied using Rimu.HamiltonianProduct. This strategy does not support adding an energy shift.

RimuRealTime.ProductSingleState — Type

ProductSingleState(v, wm, id, hamiltonian, time_step, order) <: QDSingleState

Struct holding state vector and other vectors required for Product time evolution. See QDReplicaState.

RimuRealTime.QDAlgorithm — Type

QDAlgorithm

Abstract type for quantum dynamics algorithms, for use with QuantumDynamicsProblem. Implemented algorithms: DiscretizedEvolution.

RimuRealTime.QDReplicaState — Type

QDReplicaState <: AbstractVector{QDSingleState}

Holds information about multiple replicas of QDSingleStates.

Fields

single_states: Tuple of QDSingleStates.

time_step_parameters: Time step and parameters for updating it.

shift: Energy shift.

hamiltonian: Hamiltonian.

algorithm: Algorithm.

step::Ref{Int}: Current step of the simulation

simulation_plan: Simulation plan

reporting_strategy: Reporting strategy

post_step_strategy: Post-step strategy

replica_strategy: Replica strategy

RimuRealTime.QDSimulation — Type

QDSimulation

Holds the state and the results of a quantum dynamics simulation. Is returned by init(::QuantumDynamicsProblem) and solved with solve!(::QDSimulation).

Obtain the results of a simulation sm as a DataFrame with DataFrame(sm).

Fields

problem::QuantumDynamicsProblem: The problem that was solved

state::QDReplicaState: The current state of the simulation

report::Report: The report of the simulation

modified::Bool: Whether the simulation has been modified

aborted::Bool: Whether the simulation has been aborted

success::Bool: Whether the simulation has been completed successfully

message::String: A message about the simulation status

elapsed_time::Float64: The wall time elapsed during the simulation

See also QuantumDynamicsProblem, init, solve!.

RimuRealTime.QDSimulationPlan — Type

QDSimulationPlan(; starting_step = 0, last_step = Inf, wall_time = Inf, maximum_time = 1.0)

Defines the duration of the simulation. The simulation ends when the last_step is reached, wall_time is exceeded, or maximum_time (length of real time evolution) is reached.

See QuantumDynamicsProblem, QDSimulation.

RimuRealTime.QDSingleState — Type

QDSingleState

Abstract type for single states for use with different EvolutionStrategys. QDReplicaState holds the Hamiltonian and time step information.

Concrete types:

PECSingleState

RKSingleState

EulerSingleState

ProductSingleState

RimuRealTime.QuantumDynamicsProblem — Type

QuantumDynamicsProblem(hamiltonian; kwargs...)

Defines a problem for time evolution under the given hamiltonian.

Keyword arguments:

time_step = 0.01: Size of the time step. For complex time steps, this is the modulus; use keyword alpha to define the argument.

maximum_time = 1.0: How long to evolve for in real time. Alternatively, set last_step to limit the number of time steps.

shift = 0.0: Energy shift applied to the Hamiltonian. The state evolves under H - shift*I.

initial_walkers = 1000: Initial walker population.

start_at = starting_address(hamiltonian): The initial state, as an address or an AbstractDVec.

style = IsDynamicSemistochastic{ComplexF64}(): Stochastic style of the simulation.

initiator = false: Whether to use initiators. Can be true, false, or a valid Rimu.InitiatorRule.

threading: Default is to use multithreading and/or MPI if available. Set to true to force PDVec for the starting vector, false for serial computation; may be overridden by start_at.

evolution_strategy = PEC(): Strategy for time evolution, see EvolutionStrategy.

scaling_strategy = NoScaling: Strategy for controlling walkers by scaling the vector, see ScalingStrategy.

n_replicas = 1: Number of synchronised independent simulations.

replica_strategy = NoStats(n_replicas): Which results to report from replica simulations. See Rimu.ReplicaStrategy.

reporting_strategy = ReportDFAndInfo(): How and when to report results, see ReportingStrategy in the Rimu repository.

post_step_strategy = (): Extract observables (e.g. Rimu.ProjectedEnergy), see Rimu.PostStepStrategy.

alpha = 0.0: Initial phase angle of the time step.

time_step_strategy = ConstantTimeStep(): Defines how the time step is updated during the simulation. The other implemented strategy is WalkerControl.

algorithm = DiscretizedEvolution(; time_step_strategy, evolution_strategy, scaling_strategy): The algorithm to use. Currently only DiscretizedEvolution is implemented.

starting_step = 0: Starting step of the simulation.

wall_time = Inf: Maximum time allowed for the simulation.

simulation_plan = QDSimulationPlan(; starting_step, last_step, wall_time, maximum_time): Defines the duration of the simulation. Takes precedence over last_step, maximum_time, and wall_time.

initial_time_step_parameters = (; abs_time_step, alpha, D).

display_name = "QDSimulation": Name displayed in progress bar (via ProgressLogging).

metadata: User-supplied metadata to be added to the report. Must be an iterable of pairs or a NamedTuple, e.g. metadata = ("key1" => "value1", "key2" => "value2"). All metadata is converted to strings.

random_seed = true: Provide and store a seed for the random number generator. If set to true, a new random seed is generated from RandomDevice(). If set to number, this number is used as the seed. This seed is used by solve (and init) to re-seed the default random number generator (consistently on each MPI rank) such that solveing the same QuantumDynamicsProblem twice will yield identical results. If set to false, no seed is used and consecutive random numbers are used.

RimuRealTime.RKSingleState — Type

RKSingleState(v, wm, id, hamiltonian, time_step, damping) <: QDSingleState

Struct holding state vector and other vectors required for RungeKutta time evolution. See QDReplicaState.

RimuRealTime.RungeKutta — Type

RungeKutta(damping=0) <: EvolutionStrategy

EvolutionStrategy for evolution using a second-order Runge-Kutta algorithm. In each step the state is updated according to $v_{n+1} = v_n + u_1 u_2 v_n - u_2 v_n$, where $u1 = 1 - i H dt$ and $u2 = 1 - (1 + d) i H dt / 2$, and $d$ is the damping coefficient that modifies the second-order term. Second-order damping can counteract the effects of large spectral components in the Hamiltonian that may lead to an unphysical growth of the 2-norm of the state vector.

RimuRealTime.ScalingStrategy — Type

ScalingStrategy

Abstract type for scaling strategies used to control the walker number. Passed as a parameter to QuantumDynamicsProblem or to DiscretizedEvolution. Strategies may be implemented by adding a method for scale_state_vector!.

Implemented strategies:

NoScaling

ConstantScaling

DynamicScaling

RimuRealTime.TimeStepParameters — Type

TimeStepParameters

Struct for storing the total time and parameters related to the time step.

RimuRealTime.WalkerControl — Type

WalkerControl(update_strength) <: TimeStepStrategy

Update the phase angle of the time step to control the walker number. The time step is $dt \exp(-iα)$, where $α$ is updated according to

\[α_{n+1} = α_{n} + D \arctan\left(\frac{N_\mathrm{w}^{n+1}}{N_\mathrm{w}^n}\right),\]

where $D$ is the update_strength.

The time step amplitude $dt$ and the starting angle $α$ are determined using keyword arguments time_step and alpha passed to QuantumDynamicsProblem.

CommonSolve.init — Method

init(problem::QuantumDynamicsProblem)::QDSimulation

Initialise a QDSimulation.

See also QuantumDynamicsProblem, solve!, solve, step!, QDSimulation.

CommonSolve.solve — Function

solve(::QuantumDynamicsProblem)::QDSimulation

Initialize and solve a QuantumDynamicsProblem until the maximum time is reached, the last step is completed, or the wall time limit is reached.

See also init, solve!, step!, QDSimulation.

CommonSolve.solve! — Method

solve!(sm::QDSimulation; kwargs...)::QDSimulation

Solve a QDSimulation until the maximum time is reached, the last step is completed, or the wall time limit is reached.

To continue a previously completed simulation, set a new maximum_time, last_step, or wall_time using the keyword arguments. Optionally, changes can be made to the replica_strategy, the post_step_strategy, or the reporting_strategy.

Optional keyword arguments:

maximum_time = nothing: Set the maximum time to a new value and continue the simulation.

last_step = nothing: Set the last step to a new value and continue the simulation.

wall_time = nothing: Set the allowed wall time to a new value and continue the simulation.

reset_time = false: Reset the elapsed_time counter and continue the simulation.

empty_report = false: Empty the report before continuing the simulation.

replica_strategy = nothing: Change the replica strategy. Requires the number of replicas to match the number of replicas in the simulation sm. Implies empty_report = true.

post_step_strategy = nothing: Change the post-step strategy. Implies empty_report = true.

reporting_strategy = nothing: Change the reporting strategy. Implies empty_report = true.

metadata = nothing: Add metadata to the report.

See also QuantumDynamicsProblem, init, solve, step!, QDSimulation.

CommonSolve.step! — Method

step!(sm::QDSimulation)::QDSimulation

Advance the simulation by one step.

Calling solve! will advance the simulation until the maximum time is reached, the last step is completed, or the wall time limit is reached. When completing the simulation without calling solve!, the simulation report needs to be finalised by calling Rimu.finalize_report!.

See also QuantumDynamicsProblem, init, solve!, solve, QDSimulation.

RimuRealTime.NthOrderTimeEvolution — Method

NthOrderTimeEvolution(H::AbstractHamiltonian, dt, N) <: AbstractOperator{ComplexF64}

Time evolution operator that approximates time evolution under a Hamiltonian H as the Nth order Taylor expansion of$\exp(-iHdt)$. Apply to an AbstractDVec using Rimu.apply_operator! to evolve the state. If N == -1, returns an ExponentialSampler, so the vector must not use exact spawning.

RimuRealTime.advance! — Method

advance!(report, state::QDReplicaState, s_state::QDSingleState)

Advance the state s_state by one step, and write data to the report.

RimuRealTime.clock_projector — Method

clock_projector(t) -> ClockOperator

Operator to calculate vector-vector overlaps at time step t with Clock Hamiltonians. See ClockObservable, ClockOperator.

RimuRealTime.fock_address — Method

fock_address(a::ClockAddress) -> AbstractFockAddress

Returns the address in the underlying Fock space of the ClockAddress.

RimuRealTime.num_steps — Method

num_steps(::Clock) -> Int

Return the number of time steps used in the Clock Hamiltonian.

RimuRealTime.scale_state_vector! — Method

scale_state_vector!(scaling_strategy::ScalingStrategy, state_vector, current_scale) ->
    names, stats, current_scale

Modify the state_vector according to the scaling_strategy and return the statistics to be reported, and the cumulative current_scale.

RimuRealTime.starting_state — Method

starting_state(::Clock)

Return the starting state used to construct the Clock Hamiltonian. This is an AbstractDVec in the underlying Fock space representing the state at $t=0$.

RimuRealTime.time_evolution_operator — Method

time_evolution_operator(::Clock) -> AbstractOperator

Return the time evolution operator used to construct the Clock Hamiltonian.

RimuRealTime.time_index — Method

time_index(a::ClockAddress) -> Int

Returns the time index of the ClockAddress.