Clock Hamiltonian

RimuRealTime.Clock — Type

Clock(u, length; start_at, penalty=1.0) <: AbstractHamiltonian{ComplexF64}

Clock Hamiltonian using time evolution operator u with length time steps. The operator u should be a unitary or approximately unitary AbstractOperator with a defined adjoint; for example FirstOrderTimeEvolution, NthOrderTimeEvolution, or ExponentialSampler.

The optional argument start_at specifies the state of the system at $t=0$, otherwise this defaults to DVec(starting_address(u) => 1.0). This should be an AbstractDVec in the underlying Fock space.

penalty is the size of the multiplier on the $t=0$ diagonal element that forces the initial state to remain as start_at.

The address type for the Clock Hamiltonian is ClockAddress - these addresses are built from addresses in the underlying Fock space and an additional time index, which runs from 0 to length.

Information about the clock is accessed with the following:

time_evolution_operator(clock) - the time evolution operator u

num_steps(clock) - the number of time steps length

starting_state(clock) - the starting state start_at

For calculating observables with Clock Hamiltonians, see ClockObservable, ClockOperator, clock_projector.

RimuRealTime.ClockAddress — Type

ClockAddress(address, t) <: AbstractFockAddress

Address type for use with Clocks. Stores an address in the underlying Fock space, and a time step index t. These are accessed with fock_address and time_index.

RimuRealTime.ClockObservable — Type

ClockObservable(obs::AbstractObservable, t::Int) <: AbstractObservable

Observable wrapper for use in replica strategy AllOverlaps with Clock Hamiltonians. Observable expectation values at time index t are calculated using a ClockObservable for the numerator and a clock_projector with the same time index for the denominator. Where possible, a ClockOperator is returned instead of a ClockObservable for efficiency.

RimuRealTime.ClockOperator — Type

ClockOperator(op::AbstractOperator, t::Int) <: AbstractOperator

Operator wrapper for use in replica strategy AllOverlaps with Clock Hamiltonians. Operator expectation values at time index t are calculated using a ClockOperator for the numerator and a clock_projector with the same time index for the denominator.

RimuRealTime.clock_projector — Function

clock_projector(t) -> ClockOperator

Operator to calculate vector-vector overlaps at time step t with Clock Hamiltonians. See ClockObservable, ClockOperator.

RimuRealTime.fock_address — Function

fock_address(a::ClockAddress) -> AbstractFockAddress

Returns the address in the underlying Fock space of the ClockAddress.

RimuRealTime.num_steps — Function

num_steps(::Clock) -> Int

Return the number of time steps used in the Clock Hamiltonian.

RimuRealTime.starting_state — Function

starting_state(::Clock)

Return the starting state used to construct the Clock Hamiltonian. This is an AbstractDVec in the underlying Fock space representing the state at $t=0$.

RimuRealTime.time_evolution_operator — Function

time_evolution_operator(::Clock) -> AbstractOperator

Return the time evolution operator used to construct the Clock Hamiltonian.

RimuRealTime.time_index — Function

time_index(a::ClockAddress) -> Int

Returns the time index of the ClockAddress.

Index

RimuRealTime.Clock

RimuRealTime.ClockAddress

RimuRealTime.ClockObservable

RimuRealTime.ClockOperator

RimuRealTime.clock_projector

RimuRealTime.fock_address

RimuRealTime.num_steps

RimuRealTime.starting_state

RimuRealTime.time_evolution_operator

RimuRealTime.time_index