Population Control

RimuRealTime.ScalingStrategy — Type

ScalingStrategy

Abstract type for scaling strategies used to control the walker number. Passed as a parameter to QuantumDynamicsProblem or to DiscretizedEvolution. Strategies may be implemented by adding a method for scale_state_vector!.

Implemented strategies:

NoScaling

ConstantScaling

DynamicScaling

RimuRealTime.NoScaling — Type

NoScaling() <: ScalingStrategy

Default ScalingStrategy that does not scale the vector.

RimuRealTime.ConstantScaling — Type

ConstantScaling(scale) <: ScalingStrategy

Scale the vector by $\exp(-scale*dt)$ every step. The exponential is approximated by the EvolutionStrategy.

RimuRealTime.DynamicScaling — Type

DynamicScaling(target_walkers) <: ScalingStrategy

Scale the vector every step so that the 1-norm is equal to target_walkers. The cumulative scale is stored in the DataFrame as scale.

RimuRealTime.WalkerControl — Type

WalkerControl(update_strength) <: TimeStepStrategy

Update the phase angle of the time step to control the walker number. The time step is $dt \exp(-iα)$, where $α$ is updated according to

\[α_{n+1} = α_{n} + D \arctan\left(\frac{N_\mathrm{w}^{n+1}}{N_\mathrm{w}^n}\right),\]

where $D$ is the update_strength.

The time step amplitude $dt$ and the starting angle $α$ are determined using keyword arguments time_step and alpha passed to QuantumDynamicsProblem.

RimuRealTime.scale_state_vector! — Function

scale_state_vector!(scaling_strategy::ScalingStrategy, state_vector, current_scale) ->
    names, stats, current_scale

Modify the state_vector according to the scaling_strategy and return the statistics to be reported, and the cumulative current_scale.

Index

RimuRealTime.ConstantScaling

RimuRealTime.DynamicScaling

RimuRealTime.NoScaling

RimuRealTime.ScalingStrategy

RimuRealTime.WalkerControl

RimuRealTime.scale_state_vector!