Real-Time Dynamics

Core Types

RimuRealTime.RimuRealTime — Module

RimuRealTime

RimuRealTime is a package for simulating many-body quantum systems in real time.

Welcome to RimuRealTime version 0.1.0-DEV !

RimuRealTime.QuantumDynamicsProblem — Type

QuantumDynamicsProblem(hamiltonian; kwargs...)

Defines a problem for time evolution under the given hamiltonian.

Keyword arguments:

time_step = 0.01: Size of the time step. For complex time steps, this is the modulus; use keyword alpha to define the argument.

maximum_time = 1.0: How long to evolve for in real time. Alternatively, set last_step to limit the number of time steps.

shift = 0.0: Energy shift applied to the Hamiltonian. The state evolves under H - shift*I.

initial_walkers = 1000: Initial walker population.

start_at = starting_address(hamiltonian): The initial state, as an address or an AbstractDVec.

style = IsDynamicSemistochastic{ComplexF64}(): Stochastic style of the simulation.

initiator = false: Whether to use initiators. Can be true, false, or a valid Rimu.InitiatorRule.

threading: Default is to use multithreading and/or MPI if available. Set to true to force PDVec for the starting vector, false for serial computation; may be overridden by start_at.

evolution_strategy = PEC(): Strategy for time evolution, see EvolutionStrategy.

scaling_strategy = NoScaling: Strategy for controlling walkers by scaling the vector, see ScalingStrategy.

n_replicas = 1: Number of synchronised independent simulations.

replica_strategy = NoStats(n_replicas): Which results to report from replica simulations. See Rimu.ReplicaStrategy.

reporting_strategy = ReportDFAndInfo(): How and when to report results, see ReportingStrategy in the Rimu repository.

post_step_strategy = (): Extract observables (e.g. Rimu.ProjectedEnergy), see Rimu.PostStepStrategy.

alpha = 0.0: Initial phase angle of the time step.

time_step_strategy = ConstantTimeStep(): Defines how the time step is updated during the simulation. The other implemented strategy is WalkerControl.

algorithm = DiscretizedEvolution(; time_step_strategy, evolution_strategy, scaling_strategy): The algorithm to use. Currently only DiscretizedEvolution is implemented.

starting_step = 0: Starting step of the simulation.

wall_time = Inf: Maximum time allowed for the simulation.

simulation_plan = QDSimulationPlan(; starting_step, last_step, wall_time, maximum_time): Defines the duration of the simulation. Takes precedence over last_step, maximum_time, and wall_time.

initial_time_step_parameters = (; abs_time_step, alpha, D).

display_name = "QDSimulation": Name displayed in progress bar (via ProgressLogging).

metadata: User-supplied metadata to be added to the report. Must be an iterable of pairs or a NamedTuple, e.g. metadata = ("key1" => "value1", "key2" => "value2"). All metadata is converted to strings.

random_seed = true: Provide and store a seed for the random number generator. If set to true, a new random seed is generated from RandomDevice(). If set to number, this number is used as the seed. This seed is used by solve (and init) to re-seed the default random number generator (consistently on each MPI rank) such that solveing the same QuantumDynamicsProblem twice will yield identical results. If set to false, no seed is used and consecutive random numbers are used.

RimuRealTime.QDSimulation — Type

QDSimulation

Holds the state and the results of a quantum dynamics simulation. Is returned by init(::QuantumDynamicsProblem) and solved with solve!(::QDSimulation).

Obtain the results of a simulation sm as a DataFrame with DataFrame(sm).

Fields

problem::QuantumDynamicsProblem: The problem that was solved

state::QDReplicaState: The current state of the simulation

report::Report: The report of the simulation

modified::Bool: Whether the simulation has been modified

aborted::Bool: Whether the simulation has been aborted

success::Bool: Whether the simulation has been completed successfully

message::String: A message about the simulation status

elapsed_time::Float64: The wall time elapsed during the simulation

See also QuantumDynamicsProblem, init, solve!.

RimuRealTime.QDSimulationPlan — Type

QDSimulationPlan(; starting_step = 0, last_step = Inf, wall_time = Inf, maximum_time = 1.0)

Defines the duration of the simulation. The simulation ends when the last_step is reached, wall_time is exceeded, or maximum_time (length of real time evolution) is reached.

See QuantumDynamicsProblem, QDSimulation.

Solver Interface

CommonSolve.init — Function

init(problem::QuantumDynamicsProblem)::QDSimulation

Initialise a QDSimulation.

See also QuantumDynamicsProblem, solve!, solve, step!, QDSimulation.

CommonSolve.solve — Function

solve(::QuantumDynamicsProblem)::QDSimulation

Initialize and solve a QuantumDynamicsProblem until the maximum time is reached, the last step is completed, or the wall time limit is reached.

See also init, solve!, step!, QDSimulation.

CommonSolve.solve! — Function

solve!(sm::QDSimulation; kwargs...)::QDSimulation

Solve a QDSimulation until the maximum time is reached, the last step is completed, or the wall time limit is reached.

To continue a previously completed simulation, set a new maximum_time, last_step, or wall_time using the keyword arguments. Optionally, changes can be made to the replica_strategy, the post_step_strategy, or the reporting_strategy.

Optional keyword arguments:

maximum_time = nothing: Set the maximum time to a new value and continue the simulation.

last_step = nothing: Set the last step to a new value and continue the simulation.

wall_time = nothing: Set the allowed wall time to a new value and continue the simulation.

reset_time = false: Reset the elapsed_time counter and continue the simulation.

empty_report = false: Empty the report before continuing the simulation.

replica_strategy = nothing: Change the replica strategy. Requires the number of replicas to match the number of replicas in the simulation sm. Implies empty_report = true.

post_step_strategy = nothing: Change the post-step strategy. Implies empty_report = true.

reporting_strategy = nothing: Change the reporting strategy. Implies empty_report = true.

metadata = nothing: Add metadata to the report.

See also QuantumDynamicsProblem, init, solve, step!, QDSimulation.

CommonSolve.step! — Function

step!(sm::QDSimulation)::QDSimulation

Advance the simulation by one step.

Calling solve! will advance the simulation until the maximum time is reached, the last step is completed, or the wall time limit is reached. When completing the simulation without calling solve!, the simulation report needs to be finalised by calling Rimu.finalize_report!.

See also QuantumDynamicsProblem, init, solve!, solve, QDSimulation.

Index

RimuRealTime.RimuRealTime

RimuRealTime.QDSimulation

RimuRealTime.QDSimulationPlan

RimuRealTime.QuantumDynamicsProblem

CommonSolve.init

CommonSolve.solve

CommonSolve.solve!

CommonSolve.step!